+++ nvim/lua/custom/plugins/conform.lua
@@ -33,11 +33,11 @@ return {
     end,
     formatters_by_ft = {
       lua = { 'stylua' },
-      -- Conform can also run multiple formatters sequentially
-      python = { "ruff_format", "ruff_fix", "ruff_organize_imports"},
-      --
-      -- You can use 'stop_after_first' to run the first available formatter from the list
-      -- javascript = { "prettierd", "prettier", stop_after_first = true },
+      yaml = { 'prettierd' },
+      sql = { 'sqlfmt' },
+      markdown = { 'prettierd' },
+      python = { 'ruff_format', 'ruff_fix', 'ruff_organize_imports' },
+      javascript = { 'prettierd' },
     },
   },
 }